apps/laguerison/backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
